Extract Collatz sequence
