+++ src/main/java/net/projecteuler/problems/Problem014.java
@@ -1,11 +1,11 @@
 package net.projecteuler.problems;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+
+import static net.projecteuler.util.CollatzSequence.collatzSequenceForNumber;
 
 public class Problem014 {
 
@@ -18,18 +18,5 @@ public class Problem014 {
       .max(Map.Entry.comparingByValue())
       .orElseThrow()
       .getKey();
-  }
-
-  public List<Long> collatzSequenceForNumber(long number) {
-    List<Long> sequence = new ArrayList<>();
-    sequence.add(number);
-    while (number != 1) {
-      if (number % 2 == 0) {
-        sequence.add(number /= 2);
-      } else {
-        sequence.add(number = 3 * number + 1);
-      }
-    }
-    return sequence;
   }
 }
